Add round to aggregation pipeline

+++ src/sensor.backend/sensor.Application/Contracts/ReadingsDto.cs
@@ -2,9 +2,9 @@
 {
     public class ReadingsDto
     {
-        public int[] Values { get; init; }
-        public int MaxValue { get; init; }
-        public int MinValue { get; init; }
+        public double[] Values { get; init; }
+        public double MaxValue { get; init; }
+        public double MinValue { get; init; }
         public string Type { get; init; }
         public int SensorId { get; init; }
         public DateTime[] Timestamps { get; init; }
