+++ src/sensor.backend/sensor.api/Messaging/Processing/SensorMessage.cs
@@ -2,5 +2,5 @@
 {
     public int SensorId { get; init; }
     public string Type { get; init; } = string.Empty;
-    public float Value { get; init; }
+    public double Value { get; init; }
 }
